+++ neo/app/model/SchemaForm.scala
@@ -1,0 +1,5 @@
+package model
+
+class SchemaForm {
+
+}
