+++ neo/app/views/textarea.scala.html
@@ -1,0 +1,15 @@
+@(field: Field, rows: String = "3", label: String = "CHANGEME", placeholder: String = "", help: String = "")
+
+  <div class="form-group @if(field.hasErrors) {has-error}">
+    <label class="col-sm-2 control-label">@label</label>
+    <div class="col-sm-10">
+      <textarea class="form-control"
+                rows="@rows" 
+                id="@field.id" 
+                name="@field.name" 
+                placeholder="@placeholder" 
+                >@field.value.getOrElse("")</textarea>
+      <span class="help-block">@help</span>
+      <span class="help-block">@{field.error.map { error => error.message }}</span>
+    </div>
+  </div>
