+++ neo/conf/application.conf
@@ -96,4 +96,4 @@ asterixdb.lang = SQLPP
 us.city.path = "/public/data/city.sample.json"
 
 asterixdb.view.meta.name = "viewMeta"
-tweet.url = "http://twitter-search-proxy.herokuapp.com/search/tweets?q=%s"
+tweet.url = "http://twitterMapIndex-search-proxy.herokuapp.com/search/tweets?q=%s"
