+++ neo/conf/production.conf
@@ -80,4 +80,4 @@ asterixdb.url = "http://actinium.ics.uci.edu:19002/query/service"
 asterixdb.lang = SQLPP
 us.city.path="/public/data/city.json"
 asterixdb.view.meta.name = "viewMeta"
-tweet.url = "http://twitter-search-proxy.herokuapp.com/search/tweets?q=%s"
+tweet.url = "http://twitterMapIndex-search-proxy.herokuapp.com/search/tweets?q=%s"
