+++ neo/public/stylesheets/cloudberry.css
@@ -1,0 +1,77 @@
+/*
+* multi-line comment
+*/
+p{ line-height: 1em; }
+h1, h2, h3, h4{
+    color: orange;
+    font-weight: normal;
+    line-height: 1.1em;
+    margin: 0 0 .5em 0;
+}
+h1{ font-size: 1.7em; }
+h2{ font-size: 1.5em; }
+a{
+    color: black;
+    text-decoration: none;
+}
+a:hover,
+a:active{ text-decoration: underline; }
+
+/* you can structure your code's white space so that it is as readable for when you come back in the future or for other people to read and edit quickly */
+
+body{
+    font-family: arial; font-size: 80%; line-height: 1.2em; width: 100%; margin: 0; background: #eee;
+}
+/* you can put your code all in one line like above */
+#page{ margin: 20px; }
+
+/* or on different lines like below */
+#logo{
+    width: 35%;
+    margin-top: 5px;
+    font-family: georgia;
+    display: inline-block;
+}
+/* but try and be as concise as possible */
+#nav{
+    width: 60%;
+    display: inline-block;
+    text-align: right;
+    float: right;
+}
+#nav ul{}
+#nav ul li{
+    display: inline-block;
+    height: 62px;
+}
+#nav ul li a{
+    padding: 20px;
+    background: orange;
+    color: white;
+}
+#nav ul li a:hover{
+    background-color: #ffb424;
+    box-shadow: 0px 1px 1px #666;
+}
+#nav ul li a:active{ background-color: #ff8f00; }
+
+#content{
+    margin: 30px 0;
+    background: white;
+    padding: 20px;
+    clear: both;
+}
+#footer{
+    border-bottom: 1px #ccc solid;
+    margin-bottom: 10px;
+}
+#footer p{
+    text-align: right;
+    text-transform: uppercase;
+    font-size: 80%;
+    color: grey;
+}
+
+/* multiple styles seperated by a , */
+#content,
+ul li a{ box-shadow: 0px 1px 1px #999; }
