+++ zion/src/main/scala/edu/uci/ics/cloudberry/zion/actor/DataStoreManager.scala
@@ -68,6 +68,7 @@ class DataStoreManager(metaDataset: String,
 
   def normal: Receive = {
     case AreYouReady => sender() ! true
+    case ListAllDataset => sender() ! metaData.values.toSeq
     case register: Register => registerNewDataset(sender(), register)
     case deregister: Deregister => deregisterDataSet(sender(), deregister)
     case query: Query => answerQuery(query)
@@ -320,4 +321,5 @@ object DataStoreManager {
 
   case class AppendViewAutomatic(dataset: String)
 
+  case object ListAllDataset
 }
